Changed type of Id from int to short. Everything in RSC uses a short for its Id (items, npcs, etc).

+++ src/OpenClassic.Server/Domain/IIdentifiable.cs
@@ -2,6 +2,6 @@
 {
     public interface IIdentifiable
     {
-        int Id { get; set; }
+        short Id { get; set; }
     }
 }
